Ants is now technically done
Everything mostly works except there are a couple of features that aren't there. The biggest problems are that when a colony is destroyed it leaves a void in the map and when a player wins the game just quits. There are other things missing like queens being able to create colonies or music but those aren't technically necessary.